Deep clone
Now the new method can clone the entity by id and entities from other
tables that have a field with the same name as the id from main
table and have reference to the first table. Suitable to database
that uses the same field name between tables to relate the entities.